Removed metadata type registry
This commit closes #7

The metadata type registry was an ETL concept, namely that we could keep track of each field as we parse it and try to parse it as progressively "less strongly" typed
instances. If the entire dataset were parsed and the column 'weight' could always be parsed as an int (the narrowest type), then the registry would tell you
that it was safe to cast/parse all of the "weight" attributes on a given networkx Graph as int.

This isn't necessarily a bad thing, but after an entire year, we haven't found a single person who has used it. More people are apt to just attempt to do the cast
and deal with the fallout if it fails by removing the edge than they are to ask questions about it first.

So to simplify an already complex situation, we're removing it.