This commit Closes #35.
We still don't know why the unit test `test_laplacian_embedding_elbowcut_none` fails on the MacOS Github Action instances, but not on any other MacOS system that we've tried.

This fix addresses the issue by effectively ignoring it as we can't reproduce it elsewhere, but it took steps to make the test be skipped only on Github Actions, only when using the .github/workflows/validate_on_push.yml workflow file, and only as long as the SKIP_TEST_35=true environment variable is used when running the test.

This means that developer/user testers of topologic with MacOS will still run this unit test, and we can also create a new branch to disable this flag and continue testing topologic + Github Actions on MacOS in an attempt to see when, or if, this is fixed (since it doesn't seem to be related to us so much as Github's CI instances).